requirements.txt
asgi-redis==1.4.3
asgiref==1.1.2
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
certifi==2017.11.5
channels==1.1.8
chardet==3.0.4
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
daphne==1.4.1
Django==1.11.8
django-model-utils==3.0.0
djangorestframework==3.7.7
hyperlink==17.3.1
idna==2.6
incremental==17.5.0
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
msgpack-python==0.4.8
pypiwin32==220
pytz==2017.3
redis==2.10.6
requests==2.18.4
six==1.11.0
Twisted==17.9.0
txaio==2.8.2
uritemplate==3.0.0
urllib3==1.22
zope.interface==4.4.3
